refactor(JoinModal): seperate container from component
