+++ app/components/JoinDialog/styles.js
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+export const ButtonContainer = styled.div`
+  display: flex;
+
+  & > * {
+    flex: 1;
+  }
+
+  & > * + * {
+    margin-left: 10px;
+  }
+`;
