+++ app/containers/JoinDialog/tests/index.test.js
@@ -1,7 +1,3 @@
-/**
- * Created by helge on 21.04.17.
- */
-
 import React from 'react';
 import { shallow } from 'enzyme';
 import { JoinDialog } from '../index';
@@ -21,4 +17,3 @@ describe('Join Dialog', () => {
     expect(joinDialog.nodes[0].props.children[0].props.children.props.defaultMessage).toEqual('Sorry!');
   });
 });
-
